+++ src/Raven.Studio/typescript/viewmodels/shardViewModelBase.ts
@@ -4,8 +4,6 @@ import shard from "models/resources/shard";
 import { shardingTodo } from "common/developmentHelper";
 import accessManager from "common/shell/accessManager";
 
-
-
 abstract class shardViewModelBase extends viewModelBase {
     
     protected readonly db: database;
