+++ src/Raven.Studio/wwwroot/App/views/common/sharding/shardingContext.html
@@ -1,0 +1,10 @@
+﻿<div class="flex-horizontal" data-bind="visible: showContext, if: showContext" style="position: absolute; right: 0">
+    <div class="icon-pinned" data-bind="visible: pinned"></div>
+    <div data-bind="text: contextName" class="margin-right"></div>
+    <button class="btn btn-default" data-bind="click: changeScope, visible: canChangeScope">
+        <i class="icon-config"></i>
+    </button>
+    <button class="btn btn-default" data-bind="click: exit, visible: canCloseContext"><i class="icon-cancel"></i></button>
+</div>
+<div data-bind="compose: { model: shardSelector }"></div>
+<div data-bind="compose: { model: allShardsDialog }"></div>
