+++ src/Raven.Studio/wwwroot/App/views/database/tasks/importParent.html
@@ -1,3 +1,4 @@
+<div data-bind="compose: { model: context }"></div>
 <div class="content-margin flex-vertical">
     <div>
         <ul class="nav nav-tabs" data-bind="foreach: router.navigationModel">
